move SpecralFuncInfo to its own file
