+++ torch/testing/_internal/opinfos/__init__.py
@@ -1,5 +1,6 @@
 from .core import OpInfo, SkipInfo, DecorateInfo, _NOTHING, S, M, L
 from .db import op_db, unary_ufuncs, spectral_funcs, sparse_unary_ufuncs, shape_funcs
 from .unary import UnaryUfuncInfo
+from .spectral import SpectralFuncInfo
 # these should be moved from db.py to their own file
-from .db import SpectralFuncInfo, ShapeFuncInfo, ForeachFuncInfo
+from .db import ShapeFuncInfo, ForeachFuncInfo
